refactor: 출력하는 사다리 형태를 만드는 클래스와 메서드명 수정 및 중복 제거
